Reorganize input_adapter package
