+++ shortcut_composer/input_adapter/action_manager_utils/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: © 2022-2023 Wojciech Trybus <wojtryb@gmail.com>
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+"""Utils used by core ActionManager."""
+
+from .api_krita import Krita
+from .release_key_event_filter import ReleaseKeyEventFilter
+from .shortcut_adapter import ShortcutAdapter
+
+__all__ = ["Krita", "ReleaseKeyEventFilter", "ShortcutAdapter"]
