+++ shortcut_composer/input_adapter/action_manager_utils/shortcut_adapter.py
@@ -5,8 +5,8 @@ from time import time
 
 from PyQt5.QtGui import QKeyEvent, QKeySequence
 
+from ..complex_action_interface import ComplexActionInterface
 from .api_krita import Krita
-from .complex_action_interface import ComplexActionInterface
 from .translator import Translator
