+++ shortcut_composer/input_adapter/action_manager_utils/translator.py
@@ -1,4 +1,5 @@
-
+# SPDX-FileCopyrightText: © 2022-2023 Wojciech Trybus <wojtryb@gmail.com>
+# SPDX-License-Identifier: GPL-3.0-or-later
 
 class Translator:
     """Substitutes non-latin signs with their QWERTY equivalents."""
